refactored confirm dialogs
