+++ frontend/src/views/admin/ddahs/email-ddah-confirmation.tsx
@@ -1,64 +1,9 @@
 import { Button, Modal } from "react-bootstrap";
 import React from "react";
 import { ddahIssues, getReadableStatus } from "../ddah-table";
-import { generateHeaderCell } from "../../../components/table-utils";
 import { AdvancedFilterTable } from "../../../components/filter-table/advanced-filter-table";
 import { Ddah } from "../../../api/defs/types";
-
-interface ConfirmationDdahRowData {
-    id?: number;
-    position_code: string;
-    last_name: string;
-    first_name: string;
-    total_hours: number | null;
-    status: string;
-    issue: string;
-}
-
-const emailDdahModalColumn = [
-    {
-        Header: generateHeaderCell("Position"),
-        accessor: "position_code",
-        width: 200,
-    },
-    {
-        Header: generateHeaderCell("Last Name"),
-        accessor: "last_name",
-        maxWidth: 120,
-    },
-    {
-        Header: generateHeaderCell("First Name"),
-        accessor: "first_name",
-        maxWidth: 120,
-    },
-    {
-        Header: generateHeaderCell("Status"),
-        accessor: "status",
-        maxWidth: 100,
-    },
-    {
-        Header: generateHeaderCell("Issues"),
-        accessor: "issue",
-        width: 250,
-    },
-];
-
-function compareString(str1: string, str2: string) {
-    if (str1 > str2) {
-        return 1;
-    } else if (str1 < str2) {
-        return -1;
-    }
-    return 0;
-}
-
-function compareDDAH(d1: ConfirmationDdahRowData, d2: ConfirmationDdahRowData) {
-    return (
-        compareString(d1.position_code, d2.position_code) ||
-        compareString(d1.last_name, d2.last_name) ||
-        compareString(d1.first_name, d2.first_name)
-    );
-}
+import { ConfirmationDdahRowData, compareDDAH, ddahModalColumn } from "./utils";
 
 export function MultiEmailDdahConfirmation(props: {
     selectedDdahs: Ddah[];
@@ -109,7 +54,7 @@ export function MultiEmailDdahConfirmation(props: {
                     <AdvancedFilterTable
                         // The ReactTable types are not smart enough to know that you can use a function
                         // for Header, so we will opt out of the type system here.
-                        columns={emailDdahModalColumn as any}
+                        columns={ddahModalColumn as any}
                         data={data}
                         filterable={false}
                     />
